[HalClient] Redesign the connectInBackground
This change restricts the number of calbacks that an instance of
HalClient can add during connectInBackground call when HAL is
not connected to avoid unbounded resource usage.
It also only generate one async thread to connect to the HAL no
matter how many connecting requests are received to save resources.

Test: unit tests; on device tests
Bug: 339117199
Change-Id: I7439bce8445bec89e7236569f056f3634379a6fd