fdmq: Implement ProducerBase::size()
size() is implemented in terms of a method updateAvailable() which
calculates the available space based on consumers that are in a good
state (i.e. not overwritten). updateAvailable() also is responsible
for flagging consumers as overwritten or blocking. The goal of the
method is to combine multiple operations over the consumer list which
are likely to occur close by to each other to reduce the number of times
it is required.

NOTE: tests for size() will be added once push() is added.

Bug: 445482700
Test: atest --host chre_shmem_spmc_queue_unit_tests

Change-Id: I9dd7d6cfdda4bfd1d1a59372a8cf3ae2c1a8878d