fdmq: Implements ConsumerBase::resync()
Also implements the catch-up mechanism on overwrite. When a consumer is
overwritten, if the block list epoch has not been updated, it is safe
for the consumer to fast-forward until mOverwriteResetIndex before the
the write index.

Ensures that the consumer's cached epoch value is updated on every
successful peek()/pop() operation. Also moves the epoch from
ProducerDesc to Queue as it has no dependency on the write index moving
between blocks.

Bug: 445967147
Test: atest --host chre_shmem_spmc_queue_unit_tests

Change-Id: I8bfe7135dd417d655aa924b0f52b35c68ba44364