Clear garbage data from unset venueInfo
As of 1.12, some unused bytes of the wifi scan result were replaced with the venueInfo field. It was expected that the unused fields be set to zero, but we are now enforcing this, causing some test failures.

For pixel devices, set the venueInfo to zero at the CHPP layer when the WLAN does not indicate the capability of setting venueInfo itself.

Bug: 454319884
Test: ./load_aoc.sh, atest Pts*#BasicWifiTest
Change-Id: I3992c7b728695b9849740e4c701cad98987ba6d4