fmcq: Allows separate ConsumerDesc regions
Allows consumer descriptors to be allocated in specific regions outside
a queue's primary region and separate from the regions used for other
consumers. This is required for host-side consumers to have their own
separate writable mapping containing their descriptor(s) that is also
not accessible to other host-side consumers.

Adds the ability to track consumer descriptors in one or more
containers. The primary consumer list and nodes are allocated from
shared memory but are not part of the ABI to allow for expansion and
other producer-specific functionality.

Adds producer and consumer version fields as well as reserved space in
all ABI structs.

Bug: 447371515
Bug: 445482700
Test: atest --host chre_shmem_spmc_queue_unit_tests

Change-Id: If8048e62478cf6fa61fabdc3d1a83776a4f3077c