ContextHubEndpointEchoExecutor: Assert RPC service exists
Assert the RPC service exists when the nanoapp is loaded.
If the nanoapp is loaded and we are running the test (i.e.
it is not skipped based on a capabilities check), then
MessageRouter should be enabled on the device and the RPC
service should exist.

Bug: 452479391
Flag: EXEMPT CHRE change
Test: Run the EndpointEchoTest
Change-Id: I1d9b9b669ea28e3d56f9623075bc8e207b5887ab