fmcq: Implements VariableDataProducer
Factors out common functionality in `ProducerBase::advanceWriteIndex()`
to implement `ProducerBase::truncate()` which is used by
`VariableDataProducer::truncate()`.

Implements `VariableDataProducer` on top of `ProducerBase` by prefixing
all elements with a 4-byte (and 4-byte aligned) size. On `reserve()`,
`VariableDataProducer` maintains a pointer to the size field of the
current reserved element for convenient extension or truncation of an
element via subsequent `reserve()` or `truncate()`. After `push()` or
`commit()`, the write index is advanced so that the next element's size
field is 4-byte aligned.

`VariableDataConsumer` creation has been implemented so that
`VariableDataProducer::size()` can be used to test the other producer
APIs (size returns 0 if there are no active consumers as new consumers
sync to the current producer).

Bug: 455007019
Test: atest --host chre_shmem_spmc_queue_unit_tests

Change-Id: Ic0137bb3028d356b95b8efa456cb4873ff44049e