CHPP message PAL, client, service and unit test initial merge
This CL includes
- CHPP message PAL
- client & service implementation
- unit tests

The service will run in CHRE and the client will run in the FW.
The service will not register by default unless enabled by the
makefile.

When feature enabled, CHPP init will register a CHPP message hub.
When the message service is registered, it will create an endpoint
and register with the router through the hub.

Bug: 453756093
Test: unit test and presubmit
Change-Id: I3ba6d350522c2f2dc9d0e1304241ff11d4635701