Refactor CHRE AP sensor with min NDK version 23
Refactor the CHRE AP sensor processing logic with APIs only allowed
under for Android 23.

ASensorEventQueue_registerSensor is a more tricky stuff as it is claimed
to be supported on Android 23 but GMS couldn't pass linking. Use dlsym
to get this function dynamically and returns empty sensor list if not
found. This implicitly enables the support for devices that support this
function.

Test: on Pixel 9 GMS & chre_ap_sim
Bug: 445584823
Change-Id: I450a57b2472346c4f38b85cd497d7187c8e0e847