data_flow: Implements NotificationManager.
Abstracts out the epoll_wait() functionality behind an EpollWaiter
interface to allow users to supply the thread that does the wait.

Handles mappings from data flow ids and endpoint ids to the set of fds
associated with a particular data flow (and possibly endpoint). All fds
are owned by a core map of std::unique_ptr<NotificationData>. All other
mappings use the fixed NotificationData * to look up the core mapping
when required.

Notification data contains 2 or 3 fds depending on whether this
endpoint is using that data to wait on notifications. This endpoint
waits on notifications to the producer for a host producer data flow
and notifications to a consumer for an offload producer data flow. In
these cases, the 3rd fd is not used to wait but instead to acknowledge
waking notifications from the HAL. Outgoing notifications just write 1
to the appropriate eventfd.

Bug: 457453293
Test: atest --host contexthub_data_flow_unit_tests
Change-Id: Ib73124e8db3b236be80c13d33eaca0c45f14a3be