Fixes potential double free in the event loop
Since it's possible for freeEvent() to be called from multiple places,
move the refcount decrement to inside the function, and free the event
if and only if the refcount reached zero. This prevents potential double
frees when multi-threading is enabled, for example if an event was in
the event loop queue for multiple threads, and the event loops were both
stopped, triggering a freeing of events from the event loops.

Bug: 435246073
Test: atest chre_simulation_tests pass
Change-Id: I6bce3b01bcea39ff2891a7853bc1074adb4b12d8